modify policy filename checks
Currently the PO008 check compares the policy filename against the return value of Policy.getDisplayName(), and throws an error if they do not match. Policy.getDisplayName() returns the text value of the DisplayName element, if it exists. If it does not exist, then Policy.getDisplayName() returns the value of the name= attribute on the document element. This is an unfortunate implementation.

This CL modifies the behavior in this way:

* Policy.getDisplayName() now returns the text value of the DisplayName element if it exists, or null if it does not exist.

* PO008 now checks the filename-to-DisplayName match if and only if the DisplayName element exists. It does not consider the name= attribute. This change renames the plugin source file to more accurately indicate its purpose.

* new plugin PO006 checks the filename-to-name-attribute match. It does not consider the DisplayName element.

* plugin PO007 now uses policyName to check structure, and not DisplayName

Also, there are changes to the README.