chore: update versions of some packages, prune others out (#460)
* chore: clean up package.json

Update some of the npm modules. While updating some of the npm modules, the test ordering changed. This
uncovered some bugs in the tests, which assumed a specific ordering, and a
specific profile, without specifically coding for that. So this change also
brings changes in a few tests to get them to pass reliably.